QDENGINE: Fix more signedness warnings

+++ engines/qdengine/qdcore/qd_interface_dispatcher.cpp
@@ -517,7 +517,7 @@ bool qdInterfaceDispatcher::handle_event(int event_code, const char *event_data,
 	case qdInterfaceEvent::EVENT_BUTTON_STATE:
 		if (_cur_screen && event_data) {
 			Common::String str = event_data;
-			int pos = str.find("::");
+			size_t pos = str.find("::");
 			if (pos != Common::String::npos) {
 				if (qdInterfaceButton * p = dynamic_cast<qdInterfaceButton * >(_cur_screen->get_element(str.substr(0, pos).c_str()))) {
 					return p->activate_state(str.substr(pos + 2).c_str());
